Fixed #39

+++ Server API/Controllers/authController.cs
@@ -85,6 +85,11 @@ namespace Server_API.Controllers
             if (aut == null)
                 return NotFound();
 
+            // Verify token
+            string msg = AuthorizeHeader.VerifyTokenWithUserId(ActionContext, aut.user_id);
+            if (!String.IsNullOrEmpty(msg))
+                return BadRequest(msg);
+
             db.auths.Remove(aut);
             await db.SaveChangesAsync();
